Add Router model to link a IX to a router.
The router model helps the user to know where an IX is connected. A user might
want to created several routers representing devices in the network where
IX are connected.

If an IX has a configuration template and a router, they can be used to know if
the router configuration is up-to-date. This feature is based on the NAPALM
library. For now, the configuration inside the router cannot be modified. There
is a strong link between configuration templates and routers. For example, if a
user defines a router running on the Juniper JUNOS platform, he might want to
creates a configuration template using this platform syntax.

Platforms are limited for now. Only Juniper JUNOS and Cisco IOS-XR are
supported. The other platform is a placeholder to be used when the real router
platform is not supported by this tool.