Add view to list available peers for an IX.
This view is read-only for now. It allows a user to see all networks
available to peer on an IX. It can take quite some time for large IX
since they are multiple API calls which slow down the process.

This is not a production ready feature due to its slowness but it is a
step toward a better peering management automation process.

Also add the tests for the PeeringDB API related code. Just to make
sure that this code is still valid while PeeringDB behavior might
change.