Feature to peer with an AS from then IX peer list.
The user will now see a button to peer with an autonomous system in the
peer list for IX linked with a PeeringDB ID.
When clicking on the button, if the AS has not been recorded yet, it
will be and peering sessions will be added for each IP address.
Once the peering sessions are recorded the peer will not be listed
anymore in the peer list page.