Add search feature for several pages.
The goal is to provide search features for every pages that can have
long lists of objects. In this way the user will be able to search
one or more records inside the list.

This commit also introduces a new generic view to provide views based
around tables.

The django-filters library is now required (as listed in
requirements.txt)