Fix #9. Fix routers import.
The import feature for routers was broken when trying to specify the
platform for one or more records.

This commit introduces a new class CSVChoiceField that will from a
human-understandable string find the proper value to store inside the
database. A hint will be proposed on the import views using this field.

We also introduced a new column on the import views to see required
fields.