Import Peering Sessions from Router feature.
If a router with a valid platform is attached to an IX, it is possible
to import the peering sessions configured on this router.

To do so, a NAPALM connection is established on the router and
retrieves all BGP sessions. These sessions are then filtered using the
known prefixes used to connect to the IX. Therefore this feature also
need the IX to be linked to a PeeringDB record so that prefixes can be
gathered from there too.

This feature hasn't been tested yet, so use it at your own risks.