Do not create a peering session if no remote AS.
It seems that sometimes NAPALM returns BGP peers without giving their
remote autonomous systems. It's probably a bug and an issue has been
created (napalm-automation/napalm#659).

Waiting to see if it is indeed an issue in the JUNOS module of NAPALM a
(temporary) fix is now submitted.

Import peering sessions from router feature is now tested and seems to
be working as expected. Close #8.