Implement new bulk delete feature for objects.
This feature will help user to delete several objects of the same type in only
one time. A confirmation will be ask before the actual deletion and will
summarize what is going to be deleted.

This commit also include a fix for the 'Import from Router' feature. It
actually fix a bug of display.

Get rid of objects' ID references to use PK (primary key) instead which is more
elegant in a Django project.