Check if IP version is the same when importing.
Add a conditional check to test if the IP version of the peering
session's address is the same than the current prefix of the IX. If the
versions are not the same, we should just ignore the current iteration.

Add test case to validate the behavior of the import method while
trying to import a peering session with an address which IP version
differs from the IX's prefix's.