Fix sessions not being imported. Closes #13.
While importing peering sessions from a router some of them could be
ignored due to a matching condition (session already exists). The
static method PeeringSession.does_exist() has been modified to take
positional arguments. In this way, this method is able to do more
accurate filtering to return the wanted object to check if it already
exists.

The peering sessions importation from a router now specify the IX that
it is working on when looking if a session already exists. A test case
has been added to validate that behavior.