More accurate message when config is un mergeable.
Invite the user to check the NAPALM logs file to understand why the
configuration is not mergeable. For example the LockError reported by
NAPALM in case the candidate configuration is already locked is not a
raised exception by NAPALM but just an error message. Therefore we
cannot catch the exception to inform the user about it.

Closes #16 for now.