Only show last 5 PeeringDB synchronizations.
Add hint to the "Ping" button on the router details page. This button
does not perform an ICMP test but a NAPALM connection test. By the way
this test has been improved to check if the connection is actually
alive (and usable) after opening it.

Also move the "Communities" link to the "Deployment" dropdown menu.