Add new table's column type ActionsColumn.
An ActionsColumn is a column used to display items intended to perform
actions on a given object (edit it for example). This column is only
displayed when the user has the permissions to perform said actions on
the object.

Use LinkColumn to be able to navigate to an object details page and
remove the previous button that was meant to do this.