Show peering session BGP states.
Add a new feature to retrieve and display the BGP state of peering
sessions. This feature introduces a new field for Internet exchanges.
The 'Check For Peering Session States' field must be set to 'Yes' for
states to be retrieved. The user can manually ask for an update using
the dedicated button on the peering session list view. A cron task can
also be scheduled using the new 'update_peering_session_states'
command.

Note that this feature uses a specific function of NAPALM which is not
available on every platforms. For now only JUNOS, IOS-XR and EOS are
supported.