Add password field to Peering Session. Closes #21.
Add ability to set password for peering sessions. A password can be a
clear text string or an encrypted one. However the strings are stored
as clear text in the database. It is up to the user to set the password
in the proper way so it can be used in configuration templates.

Change help text styling in forms.