Add LOGIN_REQUIRED setting to force user to log in.
Add a new middleware to make sure that the user is properly logged in before
allowing him to access any views. If the user is not logged in, he will be
redirected to the login page.

The default value for the LOGIN_REQUIRED setting is false (allowing access to
unauthenticated users).

Closes #29.