Preliminary support for NetBox devices.
Add a field to the Router model to store NetBox device ID. To enable the
support for this feature, the NETBOX_API and NETBOX_API_TOKEN must be set.

For now, it does not do much but we should consider using the management
details set in NetBox to connect to the router using NAPALM. Also we should
try to keep in sync (at least the name field) with the one found in NetBox.