Add routing policies feature.
This feature helps documenting routing policies that can then be used
in templates. Routing policies are just named objects that can be used
on routes export or routes import. It is not possible to declare the
routing policies logics.

Routing policies can be applied on an Internet Exchange or a Peering
Session level. Only one policy for export and one for import can be
defined.