Add error message when in conflict with PeeringDB.
When a PeeringDB record is deleted, if an IX is referencing to it,
there will be some issues when trying to import sessions to the IX.
This is not fundamentally wrong (the IX is supposed to be deleted) but
the user interface did not warn about it.

Starting from now an error message will be displayed with hints so the
user can be aware of the issue and take actions to fix it.

Closes #68.