Record object changes using signals.
Use Django built-in signals to record changes done to objects instead
of having to care of logging each time some code is modified.

Recording a change to an object will save the type of the object, the
type of change (created/updated/deleted), the user who's done it, and
some other details. A JSON representation is also saved for the user to
see the field values for a given object.

Closes #39.