Closes #100: Add `weight` to routing policies.
The `weight` field can be used to prioritize a routing policy over
another one. This is helpful on platform such as JUNOS on which the
policies order is important.

Routing policies are automatically sorted from the highest weight to
the lowest (in templates, tables and API). The highest weight means a
more important priority.