Allow user to add args to the BGPQ3 call.
By default the args will be set to find IPv6 routes with a prefix
length between 16 and 48 and IPv4 routes with a prefix length between 8
and 24.

The `BGPQ3_ARGS` setting must be a dictionary with two keys: `ipv6` and
`ipv4`. Each value must be a list of strings to pass to BGPQ3. No
spaces are allowed inside strings. If a string has a space in it, it
should be split into to strings.