Closes #124: Optionally use NetBox API for router access
Adds a `use_netbox` flag to the Router model. This is used to switch between
the `get_napalm` and `get_netbox` methods for pulling information from the
router.

Adds `get_bgp_neighbors` and `get_bgp_neighbors_detail` which perform the
decision to use NAPALM or NetBox.

Reworks `netbox.api` to use `pynetbox`, which handles pagination in the API
response. This rework also moved the `NETBOX_DEVICE_ROLES` filtering in to the
API request, so we're only receiving the devices we want, rather than fetching
all of them and then filtering locally. (Brought up in #32)