Do more liberal in what is accepted as an AS-SET
While people usually use AS-xxxx for simple macros, when using
hierarchical macros, some are not always using "AS-" as a prefix for
each part. For example <https://www.peeringdb.com/net/2273> is using
AS16211:RS-ANNOUNCED which is invalid. However, there is little gain
of not accepting this macro as this makes the final configuration more
secure.

This commit removes the extra check about having AS- in the macro
name.