Use ASxxxx as IRR set when no IRR set is provided (#179)
When a peer does not provide an AS-SET macro in their PeeringDB
record, we can use ASxxxxx directly to still be able to filter the
routes we are receiving from them. bgpq3 handles this correctly.

This change may surprise some users since they will get filters for
peers which previously didn't have filters.