Closes #134: Add `encrypted_password` to sessions
Users should enter clear text password and let Peering Manager do the
encryption if the router and its platform support it. This will avoid
encryption of already encrypted passwords when router platform changes.
Thus avoiding the loss of the original password string.

Passwords for sessions will be encrypted on save if the routers to
which the sessions are attached to are marked as encrypting passwords
(checkbox each router properties).

If a session is detached from a router and attached to another with a
different platform, the encrypted password will be decrypted and then
re-encrypted to match the new router platform.

It is highly recommended for users using the previous password field
with already encrypted strings to use the new mechanism. To do so, they
should decrypt current passwords (if required), store the decrypted
values in the `password` field and let the encryption in the new
`encrypted_password` happen.