Closes #185: Fix adding sessions on 2+ IXs
A session with the same IP address could not be added on more than one
IX when created from a PeeringDB record.

The IX the session is supposed to belong to is not guessed anymore, it
is passed as a value in the form submitting the sessions to be
imported.