Closes #145: Perform basic update check
If the running release is different from the latest release found on
GitHub, an yellow asterisk will be displayed next to the version
number. This icon indicates that a new release is available and links
to it. By clicking the icon the user will end up on the release page
where he'll be able to see the changelog.