Closes #209: Add option to limit prefixes per AS
Using the `--limit n` with `n` being a positive integer for the
`grab_prefixes` command will only record prefixes for an AS if their
number is lesser or equals to `n`.

This is useful to avoid recording large prefix lists that would result
in a crash due to an out of memory exception.