Add a `slug` field to `Community`
The `slug` field is meant to be used as a configuration friendly value.
Therefore the `name` field can be used however the user wants and the
`comments` field can serve a way to describe what the community is for
like a description.

Validators on the field have been turned off until proper validators
are written. Communities can be expressed differently depending on
router softwares.