Closes Router #291: Add states for routers States (Disabled/Enabled/Maintenance) (#353)
A router can now be set in different states to avoid deploying
configurations on it.

While being enabled a router will continue to receive config updates
as usual. In maintenance, configurations can also be installed but a
confirmation will be asked. When disabled, no configurations will be
sent to the router.