Closes #246: Add`net` app with `Connection` model
The `net` app will hold models and logics related to connectivity to
networks. The first model featured in this app is `Connection` which
defines objects that connect an AS to an IXP.

The `Connection` model holds details about router, IP addresses and
connectivity to Internet exchange points. As consequences the
`InternetExchange` model no longer holds these properties. It is now
used as a way to group connections thus allowing multiple
connections to the same IXP without having to duplicate objects.