Closes #476: Allow mask length in connection IPs
IP addresses with CIDR notation can be used when adding a new
connection. If the mask length is not specified it will be considered as
/32 for IPv4 and /128 for IPv6.

Add configuration context field to the connection model.

Add new `connections` Jinja2 filter that will return an iterator to loop
over the connections of a router or an IXP.