Closes #742: Send e-mail to PeeringDB network
A new view and API endpoint are provided to send an e-mail to a network
that has a known PeeringDB record. The context, compared to an AS
e-mail is different but Jinja2 filters remain the same.

Also add a new filter called `shared_facilities` to find common
facilities between two PeeringDB networks.