Implement remote data sources interactions
* Closes #365
* Closes #675

Add models and logic to synchronise some data from external
sources. It allows to use templates stored in local directories or in
git repositories. In the same way, it is also possible to push router
configurations to known data sources.

Synchronisation and push operations are performed using background tasks
and can be triggered via the user interface, the API and CLI commands.