Use cache generated value for PeeringDB sync
When responding to a query on a list endpoint, PeeringDB provides a
value corresponding to the timestamp at which the response data were
cached.

Using this value as last sync should minimise data drift over time as
the synchronisation code is pretty self healing when it comes to syncing
objects that already exist.