Closes #526: Fix available sessions list
The available sessions list was excluding sessions as long as they
appeared at least on one connection. So if an AS had more than one
connection to an IXP, a session would not appear even if it was missing
for other connections.

This fix might not be optimal from a database query perspective. It'll
generate one query per connection plus two.