Use flex for base layout
This allows to control more precisely the way we arrange the top bar,
the side bar and the main content. The missing footer, after the layout
refactoring, has been added and improved with useful links.

Also add the logo to the top left corner of the screen because I like it
this way.