Keep IRR source in AS-SET for bgpq4
As bgpq4 is able to resolve an AS-SET even if it is prefixed with the
IRR to use as source, we do not need to remove it. In fact, it's even
better to keep it so bgpq4 can do its job properly.

In the case of bgpq3, we still remove the source form the AS-SET string
and try to limit the source with it. If the parsed source is not in the
souces specified in the settings, we will query all sources.