Add journal to support non-data related logging to objects
This change implements a journal system for objects in order to be
able to log what happens to objects. This can be used to record actions
which does not imply data changes but worth mentioning for operational
reasons.

Included in this change, a journal entry will now be created when sending
an e-mail to an autonomous system.