Closes #858: Add support for AS lists
This change adds a new `as_list` field to autonomous system objects.
This field is used to store a list of ASN that can be part of the A
paths of routes advertised by an AS. This is useful to design robust
policies.

This change also deprecates the `grab_prefixes` command in favour of the
`get_irr_data` one. While they are fairly similar, the latter will cache
AS lists in addition to prefixes.

Finally, a new `as_list` Jinja2 filter is introduced to be able to use
AS lists inside templates. Using this filter will populate the AS list
for the autonomous system if it has not been cached previously.