Introduce `fields` and `exclude` API query parameters (#942)
This change has several goals.

It normalizes the default set of fields to use for all models within the
REST API. The `id`, `url`, `display_url` (when available) and `display`
fields are now always exposed in both regular and nested serializers.

It also introduces the `fields` and `exclude` query parameters that take
comma separated lists of fields that must be returned or excluded by API
serializers. This can be useful when users are interested in having just
a limited set of fields when querying the API.

Finally, the `config_context` for objects having it is now exposed. It will
provide a full rendered config context which should be more complete than
the `local_context_data` field.