Closes #943: Add endpoint to query for prefix lists for AS-SETs
This adds the required API endpoint to query for prefix lists for a list
of AS-SETs. The list can be one element or several ones (comma
separated). The values are cached and retrieved from the cache if
present, unless the `skip-cache` query parameter is passed. The cache
duration can also be changed by tweaking the `CACHE_PREFIX_LIST_TIMEOUT`
setting (defaults to 1 hour).