Add `last_used` and `allowed_ips` to tokens
The `last_used` field shows the last time a token has been used. This is
interesting to spot unused tokens.

The `allowed_ips` field is used to restrict from which IP networks a
token can be used. A network can be either a single IP address (prefix
length of /32 or /128) or a range of IP addresses by giving valid
prefixes.